feat : register prev, next로 분리
