+++ src/components/common/Layout.tsx
@@ -2,11 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
 
-interface ILayout {
-  className?: string;
-}
-
-const Layout = ({ className }: ILayout) => {
+const Layout = () => {
   const S = {
     Root: styled.div`
       background-color: #fff;
@@ -16,22 +12,10 @@ const Layout = ({ className }: ILayout) => {
       position: relative;
       border-radius: 15px;
     `,
-    App: styled.div`
-      height: 100%;
-      padding: 16px 24px;
-      &.flex-column-center {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-      }
-    `,
   };
   return (
     <S.Root>
-      <S.App className={className}>
-        <Outlet />
-      </S.App>
+      <Outlet />
     </S.Root>
   );
 };
