+++ src/pages/PaymentCardRegister.tsx
@@ -1,18 +1,7 @@
-import React, { Dispatch, useState } from 'react';
-import Title from '../components/common/Title';
-import CardPreview from '../components/CardPreview';
-import CardRegisterForm from '../components/CardForm';
-import Button from '../components/common/Button';
-import { ICardData } from '../type/CardType';
-import { Link } from 'react-router-dom';
-
-interface IContext {
-  state: ICardData;
-  dispatch: Dispatch<ICardData>;
-}
-
-export const CardDispatch = React.createContext<IContext>({} as IContext);
-
+import React, { useState } from 'react';
+import { CardDispatch } from '../context/cardDispatch';
+import PaymentCardRegisterPrev from '../components/CardRegisterPrev';
+import PaymentCardRegisterNext from '../components/CardRegisterNext';
 const PaymentCardRegister = () => {
   const [state, dispatch] = useState({
     digit1: '',
@@ -25,21 +14,16 @@ const PaymentCardRegister = () => {
     ps1: '',
     ps2: '',
   });
-  const [success, setSuccess] = useState<boolean>(false);
+  const [isComplete, setIsComplete] = useState<boolean>(false);
 
   return (
-    <>
-      <Title title={'< 카드 추가'} path={'/'} />
-
-      <CardDispatch.Provider value={{ state, dispatch }}>
-        <CardPreview />
-        <CardRegisterForm />
-      </CardDispatch.Provider>
-
-      <Link to={'/complete'}>
-        <Button text={'다음'} />
-      </Link>
-    </>
+    <CardDispatch.Provider value={{ state, dispatch }}>
+      {!isComplete ? (
+        <PaymentCardRegisterPrev setIsComplete={setIsComplete} />
+      ) : (
+        <PaymentCardRegisterNext />
+      )}
+    </CardDispatch.Provider>
   );
 };
