+++ src/stories/CardPreview.stories.tsx
@@ -7,12 +7,6 @@ export default {
   component: CardPreview,
 } as ComponentMeta<typeof CardPreview>;
 
-const Template: ComponentStory<typeof CardPreview> = (args) => (
-  <CardPreview {...args} />
-);
+const Template: ComponentStory<typeof CardPreview> = () => <CardPreview />;
 
 export const Default = Template.bind({});
-
-// Default.args = {
-//   title: '< 카드 추가',
-// };
