+++ src/styles/index.css
@@ -14,6 +14,11 @@ body {
   background-color: #e5e5e5;
 }
 
+.app{
+  height: 100%;
+  padding: 16px 24px;
+}
+
 input {
   font-size: 16px;
 }
